requirements.txt
cx_Oracle==8.3.0
kafka-python==2.0.6
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
